feat: shield working

+++ Assets/Interface/IDamageable.cs
@@ -3,5 +3,5 @@ using UnityEngine;
 public interface IDamageable
 {
     // Start is called before the first frame update
-    public void TakeDamage(int damage, GameObject causer);
+    public void TakeDamage(float damage, GameObject causer);
 }
